refactor(tests): update tests after refactoring into packages
