+++ tests/org.dataflowanalysis.analysis.tests/src/org/dataflowanalysis/analysis/tests/integration/dfd/EdgeCaseTest.java
@@ -1,4 +1,4 @@
-package org.dataflowanalysis.analysis.tests.dfd;
+package org.dataflowanalysis.analysis.tests.integration.dfd;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
