+++ tests/org.dataflowanalysis.analysis.tests/src/org/dataflowanalysis/analysis/tests/integration/dsl/IntersectionTest.java
@@ -1,4 +1,4 @@
-package org.dataflowanalysis.analysis.tests.dsl;
+package org.dataflowanalysis.analysis.tests.integration.dsl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
