+++ tests/org.dataflowanalysis.analysis.tests/src/org/dataflowanalysis/analysis/tests/integration/sequencefinder/BaseOnlineShopTest.java
@@ -28,8 +28,8 @@ public abstract class BaseOnlineShopTest {
 
     @BeforeAll
     public void initAnalysis() {
-        final var dataFlowDiagramPath = Paths.get("models", "dfd", "OnlineShop", "default.dataflowdiagram");
-        final var dataDictionaryPath = Paths.get("models", "dfd", "OnlineShop", "default.datadictionary");
+        final var dataFlowDiagramPath = Paths.get("scenarios", "dfd", "OnlineShop", "default.dataflowdiagram");
+        final var dataDictionaryPath = Paths.get("scenarios", "dfd", "OnlineShop", "default.datadictionary");
 
         analysis = new DFDDataFlowAnalysisBuilder().standalone()
                 .modelProjectName(TEST_MODEL_PROJECT_NAME)
